Move wait and effect_dom functions to sap-utils
These functions can be useful even when mocking is not required so have
been moved to the utils crate.
