+++ crates/sap-mock/Cargo.toml
@@ -7,10 +7,11 @@ edition = "2018"
 serde = "1"
 wasm-bindgen = { version = "0.2.74", features = ["serde-serialize"] }
 js-sys = "0.3.51"
-wasm-bindgen-futures = "0.4.24"
 
 [dev-dependencies]
+wasm-bindgen-futures = "0.4.24"
 wasm-bindgen-test = "0.3.24"
+sap-utils = { path = "../sap-utils" }
 
 [dev-dependencies.web-sys] 
 version = "0.3.51"
