+++ crates/sap-utils/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2018"
 
 [dependencies]
 wasm-bindgen = "0.2.74"
+js-sys = "0.3.51"
+wasm-bindgen-futures = "0.4.24"
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.24"
